refactor: 찜한 댄서 목록 조회 수정 (#59)
