+++ src/main/java/com/danthis/backend/application/dancer/DancerService.java
@@ -11,8 +11,7 @@ import com.danthis.backend.application.dancer.request.DancerAddServiceRequest;
 import com.danthis.backend.application.dancer.request.DancerUpdateServiceRequest;
 import com.danthis.backend.application.dancer.response.DancerInfoResponse;
 import com.danthis.backend.application.dancer.response.DancerSummaryListResponse;
-import com.danthis.backend.application.dancer.response.DancerSummaryResponse;
-import com.danthis.backend.application.dancer.response.PaginationInfo;
+import com.danthis.backend.application.dancer.response.DancerSummaryListResponse.DancerSummaryResponse;
 import com.danthis.backend.application.user.implement.UserReader;
 import com.danthis.backend.domain.dancer.Dancer;
 import com.danthis.backend.domain.dancer.dancerimage.DancerImage;
