+++ src/main/java/com/danthis/backend/application/dancer/response/DancerSummaryListResponse.java
@@ -1,6 +1,7 @@
 package com.danthis.backend.application.dancer.response;
 
 import java.util.List;
+import java.util.Set;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -21,5 +22,14 @@ public class DancerSummaryListResponse {
                                     .totalPages(totalPages)
                                     .totalElements(totalElements)
                                     .build();
+  }
+
+  @Getter
+  @Builder
+  public static class DancerSummaryResponse {
+
+    private Long id;
+    private String dancerName;
+    private Set<String> images;
   }
 }
